rename module `zero_balance` to `zero_ciphertext`
