+++ zk-sdk/src/sigma_proofs/zero_ciphertext.rs
